feat(ui): add generic loading, error and empty states
- extracted `EmptyStateView+Previews` to its own file
- documented static preview presets for EmptyStateView
- added preview for `ErrorMessageView` with retry support
- created `LoadingView` as reusable spinner
- updated `DiscoverView` to show loading, error and empty states
- cleaned up `DiscoverSectionView` previews: default, empty, edge cases
