+++ CineMate/CineMate/Features/Discover/Views/DiscoverSectionView.swift
@@ -10,14 +10,14 @@ import SwiftUI
 struct DiscoverSectionView: View {
     let title: String
     let movies: [Movie]
-
+    
     var body: some View {
         VStack(alignment: .leading, spacing: 8) {
             Text(title)
                 .font(.title2)
                 .bold()
                 .padding(.horizontal)
-
+            
             ScrollView(.horizontal, showsIndicators: false) {
                 HStack(spacing: 16) {
                     ForEach(movies) { movie in
@@ -38,5 +38,11 @@ struct DiscoverSectionView: View {
 #Preview("Empty") {
     DiscoverSectionView.previewEmpty
 }
-#Preview("One Movie") { DiscoverSectionView.previewOneMovie }
-#Preview("No Posters") { DiscoverSectionView.previewNoPosters }
+
+#Preview("One Movie") {
+    DiscoverSectionView.previewOneMovie
+}
+
+#Preview("No Posters") {
+    DiscoverSectionView.previewNoPosters
+}
