+++ CineMate/CineMate/Features/UI/Components/Shared/EmptyStateView.swift
@@ -52,31 +52,3 @@ struct EmptyStateView: View {
 #Preview("Generic") {
     EmptyStateView.previewGeneric
 }
-
-extension EmptyStateView {
-    static var previewNoResults: some View {
-        EmptyStateView(
-            systemImage: "magnifyingglass",
-            title: "No Results Found",
-            message: "Try adjusting your search or filter options."
-        )
-    }
-
-    static var previewNoFavorites: some View {
-        EmptyStateView(
-            systemImage: "heart.slash",
-            title: "No Favorites Yet",
-            message: "Tap the heart icon to save movies you love.",
-            actionTitle: "Browse Movies",
-            onAction: { print("Browse tapped") }
-        )
-    }
-
-    static var previewGeneric: some View {
-        EmptyStateView(
-            systemImage: "film",
-            title: "Nothing to Show",
-            message: "There’s no content here at the moment."
-        )
-    }
-}
