+++ CineMate/CineMate/Features/UI/Components/Shared/ErrorMessageView.swift
@@ -53,3 +53,4 @@ struct ErrorMessageView: View {
         onRetry: { print("Retry tapped!") }
     )
 }
+
